+++ src/main/java/com/portfolio/portfolio/repositories/SkillHardRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SkillHardRepository extends JpaRepository<SkillHardModel, Long> {
+public interface SkillHardRepository extends JpaRepository<SkillHardModel, Integer> {
 
 }
